// src/app/pages/dto/pocInformado.dto.ts
export interface IPocInformado {
    cadena_id?: string;
    poc_cadena?: string;
    poc_backus?: string;
    nombre_cadena?: string;
    nombre_backus?: string;
    direccion_backus?: string;
    gerencia_backus?: string;
    cadena_backus?: string;
    subcadena_backus?: string;
    kam_backus?: string;
    usuario_id_creacion?: string;
    usuario_id_modificacion?: string;
    fecha_tiempo_creacion?: string;
    fecha_tiempo_modificacion?: string;
  }